+++ src/views/Main/Schemas/Components/Property.js
@@ -1,6 +1,6 @@
 import React, { PropTypes } from 'react';
 import StringConsts from '../../../../utils/Consts/StringConsts';
-import { PermissionsPanel } from '../../../../components/permissions/PermissionsPanel';
+import { PermissionsPanel } from './PermissionsPanel';
 import styles from '../styles.module.css';
 
 export class Property extends React.Component {
